requirements.txt
# ============================================================
# Core Agent & Orchestration
# ============================================================
agno>=1.0.0
python-dotenv>=1.0.1
pydantic>=2.8
pyyaml>=6.0.2
rich>=13.7
tenacity>=8.5


# ============================================================
# HTTP / APIs
# ============================================================
httpx>=0.27
requests>=2.32


# ============================================================
# Data & Analytics
# ============================================================
pandas>=2.2
numpy>=2.0


# ============================================================
# GTFS-Realtime (CRITICAL)
# ============================================================
# Provides google.transit.gtfs_realtime_pb2
gtfs-realtime-bindings>=1.0.0

# Protobuf runtime (pin to safe major)
protobuf>=4.25,<6


# ============================================================
# Serialization & CLI tooling
# ============================================================
orjson>=3.10
tqdm>=4.66


# ============================================================
# Optional: SQLite persistence for agent memory
# ============================================================
# Windows uses builtin sqlite
# Linux users may need:
# pysqlite3-binary>=0.5.4 ; platform_system != "Windows"
